Specifically use stdint types instead of uint in a few locations as suggested by @rasmunk.

+++ mig/src/libnss-mig/libnss_mig.c
@@ -192,7 +192,7 @@ _nss_mig_getpwnam_r(const char *name,
         WRITELOGMESSAGE(LOG_DEBUG, "Checking for sharelink: %s\n", name);
 
         char sharelink_modes[3][11] = { SHARELINK_RW_DIR, SHARELINK_RO_DIR, SHARELINK_WO_DIR };
-        for (uint i = 0; i < 3; i++)
+        for (uint8_t i = 0; i < 3; i++)
         {
 
         memset(pathbuf, 0, PATH_BUF_LEN);
